Changed wasLastOdd variable
Changed wasLastOdd from uint256 to bool, reducing its storage size,
as well as the mapping for isOddWhitelistUser. This changed the
wasLastAddedOdd variable in the addToWhiteList function, along with
its checks. Then rearanged the state variables to better suit storage
slots. These changes reduced deployment cost and average cost of the
transferfunction and whiteTransfer